Renamed OmniShell to Shell and moved it back to the shell.py file
